feat: add instant chat initialization and refactor model selector components
ref: navigate to model route and open chat

refactor: replace SelectorLists with SelectorList component and remove unused SelectorContent file

feat: add Direct Model Access toggle in settings and update sidebar visibility

refactor: clean up Direct Model Access toggle layout in settings

fix: dropdown should change state when child selects model

fix: update model selection handling

fix: replace dispatch with modelSelectedHandler for model selection events

fix: only show tags in dropdown

refactor: remove unused selectorContentRef from Sidebar component
