+++ src/lib/components/chat/SettingsModal.svelte
@@ -70,6 +70,9 @@
 				'chat',
 				'chatbubble',
 				'chatui',
+				'directmodelaccess',
+				'direct',
+				'sidebar',
 				'username',
 				'showusername',
 				'displayusername',
