+++ src/lib/i18n/locales/de-DE/translation.json
@@ -314,6 +314,7 @@
 	"Direct Connections": "Direktverbindungen",
 	"Direct Connections allow users to connect to their own OpenAI compatible API endpoints.": "Direktverbindungen ermöglichen es Benutzern, sich mit ihren eigenen OpenAI-kompatiblen API-Endpunkten zu verbinden.",
 	"Direct Connections settings updated": "Direktverbindungs-Einstellungen aktualisiert",
+	"Direct Model Access": "Direkter Modellzugriff in Sidebar",
 	"Direct Tool Servers": "",
 	"Disabled": "Deaktiviert",
 	"Discover a function": "Entdecken Sie weitere Funktionen",
