+++ src/lib/i18n/locales/en-US/translation.json
@@ -314,6 +314,7 @@
 	"Direct Connections": "",
 	"Direct Connections allow users to connect to their own OpenAI compatible API endpoints.": "",
 	"Direct Connections settings updated": "",
+	"Direct Model Access": "Direct Model Access in Sidebar",
 	"Direct Tool Servers": "",
 	"Disabled": "",
 	"Discover a function": "",
